1. Рефакторинг тестов: улучшена структура, убраны некоторые дубликаты, повышена переиспользуемость. Также появилось явное разделение на основные группы: unit, functional, integration. Также появилось разделение на dev и test docker-compose и env. Обработаны возможные запуски интеграционных тестов не в тестовой среде. 2. Исправлены некоторые баги авторизации, автообновления access_token через middleware

+++ backend/tests/integration/conftest.py
@@ -1,3 +1,5 @@
+import os
+
 import httpx
 import pytest_asyncio
 from sqlalchemy import text
@@ -13,6 +15,10 @@ TABLES_TO_TRUNCATE = [
 
 @pytest_asyncio.fixture(loop_scope="session")
 async def clear_db():
+    env = os.environ.get("ENVIRONMENT")
+    if env != "testing":
+        raise RuntimeError(f"clear_db can only be used in testing environment. Current ENVIRONMENT={env}")
+
     async with engine.begin() as conn:
         tables = ", ".join(TABLES_TO_TRUNCATE)
         query = f"TRUNCATE TABLE {tables} RESTART IDENTITY CASCADE;"
@@ -23,6 +29,8 @@ async def clear_db():
 def user_factory():
     async def _create(client: httpx.AsyncClient, user: UserCreateDTO) -> dict:
         response = await client.post("/api/users", json=user.model_dump(mode="json"))
+        new_user = response.json()
         assert response.status_code == 200
-        return response.json()
+        assert new_user["email"] == user.email
+        return new_user
     return _create
